wip apis #2

+++ api/user/delete.php
@@ -1,5 +1,6 @@
 <?php
 //fisierul asta este pentru testare; nu ar trebui sa fie posibila stergerea asta
+// returneaza 503 daca este folosit ca foreign key deja (irelevant ^)
 header("Access-Control-Allow-Origin: *");
 header("Content-Type: application/json; charset=UTF-8");
 header("Access-Control-Allow-Methods: POST");
@@ -16,13 +17,20 @@ $user = new User($db);
 
 $data = json_decode(file_get_contents("php://input"));
 
-$user->id_utilizator = $data->id_utilizator;
-//TODO: daca nu exista user-ul
-if($user->delete()){
-    http_response_code(200);
-    echo json_encode(array("message" => "User was deleted."));
-} else {
-    http_response_code(503);
-    echo json_encode(array("message" => "Unable to delete user."));
+$user->username = $data->username;
+
+$user->getByName();
+if($user->parola === null) {
+  http_response_code(404);
+  echo json_encode(array("message" => "User does not exist."));
+}
+else {
+  if($user->delete()){
+      http_response_code(200);
+      echo json_encode(array("message" => "User was deleted."));
+  } else {
+      http_response_code(503);
+      echo json_encode(array("message" => "Unable to delete user."));
+  }
 }
 ?>
